1. Code for QuickMerge using Avro. 2. Update LinearChainWalker to properly handle cycles. 3. SimpleGraphBuilder: Helps build simple graphs for unittests. 4. Tuple class: A useful class for returning tuples of objects.
The code may not be fully compatible with versions of java earlier than 1.7. The issue is that GraphNode.clone uses Avro Builder API to copy Avro records. The Builder API uses @override to decorrate methods which are inherited from interfaces. This use of @override doesn't appear to be supported in earlier versions of java.