Several enhancements to the GraphNode API
1. Create and return hashSets of EdgeTerminals. Using hash sets is more
efficient than lists when we want to find a particular edge terminal.

2. Deprecate and replace findStrandWithEdgeToTerminal because it erroneously
assumes a node has an edge to a single strand of a node.

3. Modify removeNeighbor so we return the NeighborData that is removed from the
node. This is useful because the caller can modify the NeighborData and then add
it back. This is useful for sometimes of manipulations that require more direct
access than that exposed by NeighborData.

4. Add function addNeighbor so we can add a modified or new NeighborData to the
node.

5. Update stages for the new return type of removeNeighbor.