Fix a bug with boolean flag parsing by always requiring them to include an argument.
Previously, boolean arguments did not take a flag. Thus "--bool_arg" was
interpreted as true and the absence of the flag was false. However by not
requiring an argument we don't allow a flag to be specified as false by doing
"--bool_arg=false". This is problematic because it means we can't set a boolean
flag to true by default and then override it.