Fix bug in RemoveLowCoverage.
Suppose the low coverage node has an edge from the
forward and reverse strand to a particular node. We need to make sure a single
remove neighbor message is sent to this neighbor. Prior to this CL we would have
sent two messages, one for each strand. This would cause an error in the reduce
phase because the neighbor would try to remove the edges to the low coverage
node twice.