BuildGraph support multiple formats.
We add support so the input to the BuildGraph stage can be FastQRecord (which is
an avro schema). This allows the stage to consume the output of the error
correction stages.

We also add support to Sequence so we can read in sequences in various formats
into the current instance. This way we can reuse an existing object.