Fix issues with AssembleScaffolds getting blocked calling toAmosNew.
One issue was we need to check the error and input streams are ready before trying to read them. We also switched to polling model so while the command is running we periodically poll its output streams so that for long running commands we get periodic updates.

Scaffolding is still having errors when run on actual staph data.
The error is

toAmos_new:Error could not find eid SRR022868.8577566_2 in bank, skipping

Looking at the fasta file produced by BuildBambusInput I see the following
>SRR022868.8577566/2
AGAATTAGCAGTCATTAAATTTATC

So the problem could be that we aren't consistent in how we replace "/" with "_" in the BuildBambusInput stage.