Bug fix for cycles.
For the rhodobacter dataset, the graph was becoming corrupted by QuickMergeAvro. The problem appears to be with cycles. Here's the relevant subgraph:
https://docs.google.com/a/lewi.us/file/d/0B6ngy4MCihWwYnkzY0oyZzhRbWs/edit

Cycles are detected and handled in NodeMerge.mergeNodes. However, in QuickMergeUtil, since we merge multiple nodes without moving the edges to the start of the chain, the cycle wasn't being detected properly in NodeMerger. We fix this by detecting the cycle in QuickMergeUtil and moving the appropriate edges so that NodeMerger can detect the edges.

We added a unittest to catch this situation.

Using the new code, QuickMerge didn't produce an invalid graph for the rhodobacter dataset.