Pipelines should log information about the stages run using an avro record in json format. This will make it easy to resume execution or validate the graph after each stage.
We needed to modify the pom file to add jackson dependencies which provide
support for json encoding.

The current implementation is limited in that we create a new file after each
stage runs. Thus, we accumulate a bunch of json files. It would be good to clean
these up.

The timestamp also isn't in local time.

(Some of the changes described above might actually have gotten added in
a previous commit which added the FindLatestValid.java program)

https://codereview.appspot.com/7004052/