Major overhaul of the node merging code.
Rather than do several pairwise merges to merge a linear chain, we merge a chain
in a single pass. This makes it easier to handle cycles and other edge cases.

We fix a related bug in PairMarkAvro with cycles. In PairMarkAvro we shouldn't
be moving edges between two up nodes that are being shipped to the same down
node. Moving these edges makes it difficult to identify and handle cycles in the
NodeMerger.

https://codereview.appspot.com/7030043/