Major overhaul of the scaffolding code. We added two map reduce jobs. The first mapreduce job reads the output of bowtie and converts it to avro records. The second mapreduce joins the bowtie output with the original contigs and produces a tigr file.
These mapreduces eliminate the need to load the bowtie output into memory to do the join; an approach which doesn't scale.

In addition we fixed some issues with the mock output collectors. OutputMockCollector had the limitation that it could only handle a single output record. To fix this we modified it so that outputs are accumulated in a list. This necessitates making a copy of the objects because the caller could be reusing an object. In order to facilitate the copies we made the KEY and VALUE subclasses of Writable and used the serialization methods in Writable.

Consequently, we can no longer use OutputMockCollector for non-avro mappers which output AvroWrapper<T>. We created a separate AvroWrapperCollectorMock class for that case.