Major overhaul of the stage infrastructure. We introduce new classes 1. StageBase 2. MRStage 3. NonMRStage 4. PipelineStage
There are now specialized base classes for each class of jobs (e.g MR jobs).

More importantly, StageBase defines hooks that subclasses override. This allows the base classes to handle the overall execution flow and just gives subclasses predefined hooks to customize different parts of the execution flow. This makes it easier to ensure a consistent logging experience and removes the burden on the author of the subclass from implementing repetitive tasks.

One of the main motivations for this change was probably logging execution information StageInfo for jobs that launch other jobs. This is important so that
we can track down graph corruption bugs automatically.