Overhaul support for FastQ files in hadoop.
We replace FastQText with FastQWritable. The new class doesn't use the Text object for serializing/deserializing FastQRecords. The new implementation should be more efficient because we only parse the data once when deserializing.

FastQWritableRecordReader is a replacement for FastQRecordReader which we deleted. We should consider renaming FastQWritableRecordReader to FastQRecordReader.

We removed a lot of unused code from FastQInputFormat. We also got rid of the need to specify a maximum line length.