Remove MockFSDataInputStream.
This class was being used to be test FastQInputFormat but we were running into problems using a MockFSDataInputStream and a LineReader. In particular, LineReader.readLine() would cause an exception when trying to read from a MockFSDataInputStream. The problem appears to happen when we don't specify the maximum length of a line to read. In this case, we end up trying to read more bytes then are in the buffer backing MockFSDataInputStream and the MockFSDataInputStream doesn't seem to handle this in the same way that an FSDataInputStream does.

To avoid these problems we have already updated the tests to just write the data
to a file and then use FSDataInputStreams in our tests.