Fix a bug in QuickMergeUtil.findNodesToMerge.
This problem popped up while assembling contigs for the rhodobacter dataset.

The problem is that the code currently assumes that if we have a cycle we can
break the cycle anywhere. However, that's only if we have an imperfect cycle.
This is explained more in the comments in the code.