A bunch of changes which need to be committed together because of the changes to the stage infrastructure.
1. Update the stages in contrail.stages to use the new Stage infrastructure.
2. Fix the bug in the coverage estimate.

We were ignoring KMers coming from the middle of reads.
As part of this change we refactored the ReadState enum.
see https://codereview.appspot.com/7398048/.

3. Create a custom wrapper for loggers so we can control the behavior when we
encounter a fatal log message. i.e normaly we should exit after logging the
error. However, in testmode we should exit using "Fail" so the unittest
registers it as failure.

4. Fix a bug in QuickMergeUtil.findNodesToMerge.
This problem popped up while assembling contigs for the rhodobacter dataset.

The problem is that the code currently assumes that if we have a cycle we can
break the cycle anywhere. However, that's only if we have an imperfect cycle.
This is explained more in the comments in the code.

5. Major improvements to FindLastValidGraph. This is actually useful now.