Update CompressibleAvro so the mapper merges a node’s strands if the strands are connected. This requires passing K as a parameter to CompressibleAvro. We do this in the mapper because it is easier to handle this compression here then trying to mark the nodes as compressible and then handling the compression in PairMark/PairMerge or QuickMerge.
Refactor the code for testing if a node is compressible into a
separate function isCompressible.

Add a sanity check to isCompressible to verify the number of messages
matches the expected number in the case where we have a tail.

Add two test cases to the unittest test to verify that
CompressibleAvro properly merges connected strands of a node.

Tests:
A bunch of tests are broken.