Fix how FindBubbles and PopBubbles handle palindromes.
Palindromes should not be given special treatement. Palindromes are actually
a special case of bubbles formed by two strands of a node e.g the node A in
X->{A, R(A)}->R(X).

For now we don't try to pop the bubble formed by A and R(A). However,
we do add a counter so we can measure how often they occur.

CompressAndCorrect no longer needs to check if FindBubbles detected
palindromes. This criterion combined with the improper handling of
palindromes in FindBubbles was causing CompressionAndCorrection to
never satisfy the termination criterion and to loop endlessly.