Fix a bug in quick merge that showed up when assembling the staph dataset.
The bug was caused by a graph like:
Z->A
A->B->C->D->A
In this case QuickMerge will try to merge chain B->-C->-D.
The bug was in updating the edges in A to the merged nodes. We were modifying
A and then looking at the modified A to determine which edges need to
be added.

The solution is to update A in two phases. In the first phase we just
determine which edges to add to A but don't actually add them. Then in
the second phase we add all the edges.

We validate the fix using an integration test. The integration runs
QuickMerge on the subgraph taken from the staph dataset that was causing
problems.