Add a stage to reverse the reads for scaffolding.
The stage uses the existing ReverseReads stage but processes each file
individually so that we can preserve the original filename. This is necessary
because scaffolding use the filename to identity the library a set of reads
belongs to.

Added a method to FileHelper to match files to a comma separated list of
files or globs.