Make BuildBambusInput more efficient so that we don't run out of memory.
There's no reason to load into memory the id's of all mate pair. Rather,
we can assume mate pairs are collated in different files and verify this
as we go along. This way we can write the mate ids and shortened reads to
the output iteratively and not need to store them in memory.

Add some utilities for verifying if two reads form a mate pair based on
their ids.