Major changes to improve scaffolding.
*) Split AssembleScaffolds into separate stages for aligning reads to
the contigs and for running bambus. This allows us to filter the alignments
before assembling the scaffolds.

*) Change BuildBambusInput to use a json file containing Library records
to describe the libraries. The tool AddLibraryToFile makes it easy to
add libraries to the file. As a result, BuildBambusInput no longer needs
to organize the files into mate pairs based on the filenames.

*) Fasta and FastQ file readers should use InputStreams and hadoop paths so
they can read from non local filesystems.

*) Add functions to read/write an array of records as a json array.

*) DNAAlphabetWithNFactory add characters 'Y' and 'R' to the list of characters
treated as errors. The speciesA dataset for the assemblathon appears to
include those charactes.

*) DNAUtil: Create a separate function for computing the DNA complement of a
base. We do this so that we could potentially use this to rewrite ReverseReads
to ignore all characters other than A,C,G,T.

*) Replace AssembleScaffolds with RunBambus.

*) Create a separate stage AlignReadsWithBowtie to align the reads with bowtie.

*) Add crunch as a dependency so we can start writing crunch pipelines.

*) CrunchStage is the basis for stages which use crunch.

*) FilterBowtieAlignments filters bowtie alignments so we don't crash
Bambus.