Overhaul the scaffolding pipeline so we can filter the data after aligning the reads to the contigs before building the Bambus input. This reduces the data loaded into Bambus preventing Bambus from running out of memory for the species A dataset for assemblathon 1.
* AvroFileConentsIterator should respect the scheme in the file paths if there
 is one.

* Rewrite BuildBambusInput to use BuildBambusFastaFile and BuildBambusLibraryFile.

* Create a stage to build the fasta input file for bambus.

* BuildBambusLibraryFile a separate stage for building the mates/library
file for bambus.

* AlignReadsWithBowtie should not be shortening the reads; we already do
that in BowtieRunner. I think this was a legacy from a previous iteration
of the code when we were creating the shortened reads to import into Bambus.

* In BowtieRunner we try not using safeReadId to convert "/" to "_" because
this just causes headaches if we change the read ids.

* Create new versions of the FastQInputFormat which use the new mapreduce APIs
as opposed to the old mapred API. This way we can use the new APIs with Crunch.

* Refactor to share code between the two FastQInputFormat classes.
* Move the classes for the old mapreduce API into contrail.io.mapred.
* The bulk of the shared code is in contrail.io.FastQSplitter.
* Add an integration test for FastQRecordREader which uses crunch.

* FilterReads is a crunch stage for scaffolding to only filter the reads for which there is a
bowtie mapping. We need to do this because Bambus keeps in memory a hash table of all
read ids. If we use all reads we run out of memory with the species A dataset from
the assemblathon.

* Add functions for parsing the read ids in the asemblathon dataset.
* Add DoFns for forming mate pairs and grouping mate pairs based on the library.

* AlignReadsWithBowtie no longer changes "/" to "_" by using Utils.safeReadId.
Its unclear why we needed to do this. I think it was inherited from an earlier
version of the code. This was causing problems downstream since the read ids
no longer matched the original version.

* MatePairEdges is a Crunch stage for computing information about mate pair links
between contigs.